requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
attrs
av
Babel
beautifulsoup4
blinker
cachetools
certifi
chardet
charset-normalizer
click
colorama
colorzero
cryptography
distlib
distro
filelock
Flask
frozenlist
google-api-core
google-auth
google-cloud-texttospeech
google-genai
googleapis-common-protos
gpiozero
greenlet
grpcio
grpcio-status
h11
html5lib
httpcore
httpx
idna
importlib-metadata
itsdangerous
Jinja2
jsonpointer
jsonschema
lgpio
libevdev
lxml
MarkupSafe
meson
more-itertools
multidict
numpy
oauthlib
olefile
pexpect
pgzero
picamera2
pidng
piexif
pigpio
Pillow
platformdirs
playwright
propcache
proto-plus
protobuf
ptyprocess
pvporcupine
pyasn1
pyasn1_modules
PyAudio
pycryptodomex
pydantic
pydantic_core
pyee
pygame
pyinotify
PyJWT
PyOpenGL
pyOpenSSL
pyrsistent
pyserial
python-dotenv
python-prctl
python-weather
pytz
regex
reportlab
requests
requests-oauthlib
responses
rfc3987
rpi-lgpio
RPi.GPIO
rsa
scipy
simpleaudio
simplejpeg
simplejson
six
smbus2
sniffio
soupsieve
SpeechRecognition
spidev
tenacity
toml
tqdm
twython
typing-inspection
typing_extensions
uritemplate
urllib3
v4l2-python3
videodev2
virtualenv
volume-control
webcolors
webencodings
websockets
Werkzeug
yarl
zipp
geocoder
